chore: moving calculator e shared to new folders
