+++ src/app.module.ts
@@ -3,13 +3,13 @@ import { ConfigModule } from '@nestjs/config';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { CalculatorModule } from './calculator/calculator.module';
 import { DatabaseModule } from './database/database.module';
+import { DocumentsModule } from './documents';
 import { FormsModule } from './forms';
 import { GraphQLModule } from './graphql/graphql.module';
-import { MailModule } from './mail/mail.module';
+import { CalculatorModule } from './modules/calculator/calculator.module';
+import { MailModule } from './shared/mail/mail.module';
 import { UsersModule } from './users/users.module';
-import { DocumentsModule } from './documents';
 
 @Module({
   imports: [
@@ -25,4 +25,4 @@ import { DocumentsModule } from './documents';
   controllers: [AppController],
   providers: [AppService],
 })
-export class AppModule { }
+export class AppModule {}
