+++ src/shared/mail/interfaces/mail.interface.ts
